+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 See also [TODO](TODO.md)
 
 ## [ ] Version 1.1
-- [ ] simplify the side-by-side view `get_combined` function since now only dedicated fills are supported
+- [x] missing environment variables `EMAIL_*` do not cause server exception, but display message in django logs that email not sent
+- [x] add filter in blotter/concealed view: by order side
 - [x] run the `OrderManager.process` function at each page load
   - alternatives could have been: cron it; use django's cellery to schedule it as a repeated job